Optimize variable on Varmo

+++ varmo/V1.0.0/Varmo/Varmo.h
@@ -46,8 +46,6 @@ String "machine.alive";
 String "machme.alive:ok";
  */
 
-
-
 const uint8_t MODE_POS = 1;
 const uint8_t MODE_POS_SPD = 2;
 const uint8_t MODE_HOME = 3;
@@ -70,20 +68,20 @@ void serialEvent();
 void doEncoderA();
 void doEncoderB();
 
-int menu_set(int MENU);
+uint8_t menu_set(uint8_t MENU);
 float resolution_set(float RESOLUTION); 
-void menu_init(int MODE, int *CONTRAST, float *POSITION, float * TORQUE, float * SPEED, float *HOME_POSITION, float *ACCELERATION, float *DECELERATION, float * encoder0Pos, float resolution);
+void menu_init(uint8_t MODE, uint8_t *CONTRAST, float *POSITION, float * TORQUE, float * SPEED, float *HOME_POSITION, float *ACCELERATION, float *DECELERATION, float * encoder0Pos, float resolution);
 
-void lcd_print_menu(int *MODE, int CONTRAST, float POSITION, float TORQUE, float SPEED, float HOME_POSITION, float ACCELERATION, float DECELERATION, float POS_SPEED, 
+void lcd_print_menu(uint8_t *MODE, uint8_t CONTRAST, float POSITION, float TORQUE, float SPEED, float HOME_POSITION, float ACCELERATION, float DECELERATION, float POS_SPEED, 
 					float torque_get, float speed_get, float position_get, float home_position_get, float acceleration_get, float decelaration_get, float pos_speed_get, float *encoder0Pos);
-void lcd_print_contrast_value(int CONTRAST);
+void lcd_print_contrast_value(short CONTRAST);
 void lcd_print_sign(float value);
 void lcd_print_float_align_right(float value);
 void lcd_print_float_value(float value1, float value2);
 void lcd_print_float_value_three(float value1, float value2);
 
 void converter(float *value, float *encoder0Pos, float resolution, bool sens, float max);
-void contrast_convert(int *CONTRAST, int *F_contrast, float *encoder0Pos);
+void contrast_convert(uint8_t *CONTRAST, uint8_t *F_contrast, float *encoder0Pos);
 
 #endif // VARMO_H_INCLUDED
